double to float

+++ project/LBDefinitions.h
@@ -12,15 +12,15 @@ static const int LATTICEVELOCITIES[Q][D] = {
     {0, -1, 1}, {-1, 0, 1}, {0, 0, 1}, {1, 0, 1}, {0, 1, 1}
 };
 
-static const double LATTICEWEIGHTS[Q] = {
+static const float LATTICEWEIGHTS[Q] = {
     1.0/36, 1.0/36, 2.0/36, 1.0/36, 1.0/36,
     1.0/36, 2.0/36, 1.0/36, 2.0/36, 12.0/36, 2.0/36, 1.0/36, 2.0/36, 1.0/36, 
     1.0/36, 1.0/36, 2.0/36, 1.0/36, 1.0/36
 };
 
 /* Math.sqrt(3.0) = 1.732050807568877 */
-static const double C_S = 1.0 / 1.732050807568877;
-static const double C_S2 = 1.0 / 3.0; 
+static const float C_S = 1.0 / 1.732050807568877;
+static const float C_S2 = 1.0 / 3.0; 
 
 static const int FLUID = 0;
 static const int NOSLIP = 1;
